+++ 0.57.1/src/CCharNPCAct.cpp
@@ -2581,7 +2581,7 @@ void CChar::NPC_OnTickAction()
 
 	EXC_CATCH;
 	EXC_DEBUG_START;
-	g_Log.Debug("'%s' [0%lx]\n", GetName(), GetUID());
+	g_Log.Debug("'%s' [0%lx]\n", GetName(), (DWORD)GetUID());
 	EXC_DEBUG_END;
 }
 
@@ -2638,7 +2638,7 @@ void CChar::NPC_Pathfinding()
 
 	EXC_CATCH;
 	EXC_DEBUG_START;
-	g_Log.Debug("'%s' point '%d,%d,%d,%d' [0%lx]\n", GetName(), local.m_x, local.m_y, local.m_z, local.m_map, GetUID());
+	g_Log.Debug("'%s' point '%d,%d,%d,%d' [0%lx]\n", GetName(), local.m_x, local.m_y, local.m_z, local.m_map, (DWORD)GetUID());
 	EXC_DEBUG_END;
 }
